Moved repeated code to calibration/commons
