+++ src/calibration/arucoMarkers.py
@@ -6,7 +6,7 @@ import glob
 import pprint
 from timeit import default_timer as timer
 
-from src.calibration.commons import rotationMatrixToEulerAngles
+from src.calibration.commons import rotationMatrixToEulerAngles, getImageAndResize, calculateCoordinates
 
 # Aruco common parameters
 ARUCO_PARAMETERS = cv2.aruco.DetectorParameters_create()
@@ -20,7 +20,6 @@ RFlip[2,2] = -1.0
 
 def generateMarkerGrid(nx, ny, outputFile):
     # Create gridboard, which is a set of Aruco markers
-    # the following call gets a board of markers 5 wide X 7 tall
     gridboard = cv2.aruco.GridBoard_create(
             markersX=nx, 
             markersY=ny, 
@@ -32,10 +31,6 @@ def generateMarkerGrid(nx, ny, outputFile):
 
     cv2.imshow('Aruco grid', image)
     cv2.waitKey(0)
-
-def getImageAndResize(sourceFile, scale):
-    sourceImage = cv2.imread(sourceFile)
-    return cv2.resize(sourceImage, None, fx=scale, fy=scale)
 
 def getCorners(sourceImage):
     grayImage = cv2.cvtColor(sourceImage, cv2.COLOR_BGR2GRAY)
@@ -64,14 +59,6 @@ def getPositionVectors(sourceImage, markerLength, cameraMatrix, distCoeffs):
     idsFound = np.array([x[0] for x in idsFound])    
     return idsFound, rVecs, tVecs
 
-def calculateCoordinates(rVecs, tVecs):
-    # Converting the rVector into Euler angles
-    rMatrix = np.matrix(cv2.Rodrigues(rVecs)[0])
-    roll, pitch, yaw = rotationMatrixToEulerAngles(RFlip*rMatrix) # Flipping before converting
-    traX, traY, traZ = tVecs[0]
-
-    return [math.degrees(roll), math.degrees(pitch), math.degrees(yaw), traX, traY, traZ]
-
 def estimatePose(sourceFile, outputFile, scale, markerId, markerLength, cameraMatrix, distCoeffs):
     sourceImage = getImageAndResize(sourceFile, scale)
     
@@ -87,7 +74,7 @@ def estimatePose(sourceFile, outputFile, scale, markerId, markerLength, cameraMa
         cv2.aruco.drawAxis(sourceImage, cameraMatrix, distCoeffs, rVecs[i], tVecs[i], markerLength/2)
 
         coordNames = ['Roll: ', 'Pitch: ', 'Yaw: ', 'Tra X: ', 'Tra Y: ', 'Tra Z: ']
-        coords = calculateCoordinates(rVecs[i], tVecs[i])
+        coords = calculateCoordinates(np.reshape(rVecs[i], (3,1)), np.reshape(tVecs[i], (3,1)), RFlip)
 
         font = cv2.FONT_HERSHEY_SIMPLEX
         initialPosition = (10,100)
@@ -126,10 +113,10 @@ def exportCoordinatesToFile(sourceFile, outputFile, scale, markerId, markerLengt
 
     if indexes.size > 0:
         i = indexes[0]
-        coords = calculateCoordinates(rVecs[i], tVecs[i])
+        coords = calculateCoordinates(np.reshape(rVecs[i], (3,1)), np.reshape(tVecs[i], (3,1)), RFlip)
         coords.insert(0, markerId)
 
-        with open(outputFile, mode='a') as coordsFile:
+        with open(outputFile, mode='w') as coordsFile:
             writer = csv.writer(coordsFile, delimiter=',', quotechar='"', quoting=csv.QUOTE_MINIMAL)
             writer.writerow(coords)
     else:
