+++ txt/coords/ARUA_R_0_P_30N_Y_0-coords.csv
@@ -1,2 +1,1 @@
 6,-0.2088223669722394,-29.14720119477282,-0.5741368478991092,12.75152357805032,-4.238404395029059,57.31695121725972
-6,-0.2088223669722394,-29.14720119477282,-0.5741368478991092,12.75152357805032,-4.238404395029059,57.31695121725972
