+++ txt/coords/ARUA_R_0_P_30_Y_0-coords.csv
@@ -1,2 +1,1 @@
 6,-1.1251179837839194,29.02504890831706,-0.8867945250334763,-4.556151156329105,-3.3043479686931336,62.11273014509013
-6,-1.1251179837839194,29.02504890831706,-0.8867945250334763,-4.556151156329105,-3.3043479686931336,62.11273014509013
