+++ txt/coords/ARUA_R_0_P_45N_Y_0_00-coords.csv
@@ -1,2 +1,1 @@
 6,-3.249547376597585,-43.93121804990847,0.5497021885782613,11.31311106366899,-2.1458590464618683,45.93056436616105
-6,-3.249547376597585,-43.93121804990847,0.5497021885782613,11.31311106366899,-2.1458590464618683,45.93056436616105
