+++ txt/coords/ARUA_R_0_P_45N_Y_0_01-coords.csv
@@ -1,2 +1,1 @@
 6,-3.6303196466154053,-45.17957856296011,0.2578167351257513,12.83567935287493,-1.4676438763103261,60.23674734264318
-6,-3.6303196466154053,-45.17957856296011,0.2578167351257513,12.83567935287493,-1.4676438763103261,60.23674734264318
