+++ txt/coords/ARUA_R_0_P_45_Y_0_00-coords.csv
@@ -1,2 +1,1 @@
 6,-3.002643101227997,44.43285687462659,-3.738294799502734,-5.239364800190696,-1.8631236289452962,65.39845844612016
-6,-3.002643101227997,44.43285687462659,-3.738294799502734,-5.239364800190696,-1.8631236289452962,65.39845844612016
