+++ txt/coords/ARUA_R_0_P_45_Y_0_01-coords.csv
@@ -1,2 +1,1 @@
 6,-2.9552331115474106,46.52755901980758,-2.7614501924244776,-0.5658723204479508,-1.186445002580251,78.33386984957437
-6,-2.9552331115474106,46.52755901980758,-2.7614501924244776,-0.5658723204479508,-1.186445002580251,78.33386984957437
