+++ txt/coords/ARUA_R_0_P_60N_Y_0-coords.csv
@@ -1,2 +1,1 @@
 6,-2.068854569382498,-59.46510089675468,-0.38829357916365975,6.8507240648419225,-3.1453856132261326,35.68944096591929
-6,-2.068854569382498,-59.46510089675468,-0.38829357916365975,6.8507240648419225,-3.1453856132261326,35.68944096591929
