+++ txt/coords/ARUA_R_0_P_60_Y_0-coords.csv
@@ -1,2 +1,1 @@
 6,-0.2871141177353641,57.80628541785465,-0.018632584821974683,-4.84209021775182,-3.7805625684159896,33.94253030200431
-6,-0.2871141177353641,57.80628541785465,-0.018632584821974683,-4.84209021775182,-3.7805625684159896,33.94253030200431
