+++ txt/coords/ARUZ100-coords.csv
@@ -1,2 +1,1 @@
 6,-3.51148719665937,-1.1623374724034508,-1.7358831927516134,7.232285421651814,-3.0151200232360926,101.26848822213323
-6,-3.51148719665937,-1.1623374724034508,-1.7358831927516134,7.232285421651814,-3.0151200232360926,101.26848822213323
