+++ txt/coords/ARUZ150-coords.csv
@@ -1,2 +1,1 @@
 6,0.36471288121462647,-2.1030346782730747,-0.8543542329484118,-3.719160588367876,-3.628906506337972,152.3820381636232
-6,0.36471288121462647,-2.1030346782730747,-0.8543542329484118,-3.719160588367876,-3.628906506337972,152.3820381636232
