+++ txt/coords/ARUZ200-coords.csv
@@ -1,2 +1,1 @@
 6,0.9978190208687446,-11.321837196912503,-2.358908227380436,-1.0605552536747238,-2.7147628488026565,201.048303032248
-6,0.9978190208687446,-11.321837196912503,-2.358908227380436,-1.0605552536747238,-2.7147628488026565,201.048303032248
