+++ txt/coords/ARUZ50-coords.csv
@@ -1,2 +1,1 @@
 6,-0.16199506373808623,2.6783586195310076,-2.013312128344245,5.842023489059199,-2.8020005820233522,50.192908714014465
-6,-0.16199506373808623,2.6783586195310076,-2.013312128344245,5.842023489059199,-2.8020005820233522,50.192908714014465
