+++ txt/coords/ARUZ75-coords.csv
@@ -1,2 +1,1 @@
 6,5.324221931837778,-11.120670378096838,-2.000512400012633,7.429166082404012,-2.7917298091207625,75.60009955528307
-6,5.324221931837778,-11.120670378096838,-2.000512400012633,7.429166082404012,-2.7917298091207625,75.60009955528307
